Revise mcp_stdio_client to contact existing server
The original mcp_stdio_client tries to start and stop a local
server. I would like to use the MCP client to instead talk to
some server that is already running.  This adds support for
allowing the client to contact a server at a given port and host.

Includes example code as well.